Add files via upload
The interface, implementation and main file of the linked list class. Some methods such as divideList and mergeList are not fully functional. This was my first attempt of building various operations of a linked list, thus it is not specialised in the sense that it is not a stack or queue.